==> 00b_merge_spotify_chartdata.ipynb <==
import pandas as pd
import os 
from tqdm.notebook import tqdm

country_charts = []

for file in os.listdir(os.path.join('data','chart_data')):
    if file.endswith(".csv"):
        country_charts.append(file)
        
country_charts2 = []

for file in os.listdir(os.path.join('data','chart_data','data_202106_202111')):
    if file.endswith(".csv"):
        country_charts2.append(file)

country_charts3 = []

for file in os.listdir(os.path.join('data','chart_data','data_202111_202202')):
    if file.endswith(".csv"):
        country_charts3.append(file)

dfs = []
for chart in tqdm(country_charts):
    df_ = pd.read_csv(os.path.join('data','chart_data',chart), parse_dates=['Date'])
    dfs.append(df_)

for chart in tqdm(country_charts2):
    df_ = pd.read_csv(os.path.join('data','chart_data', 'data_202106_202111', chart), 
                      parse_dates=['Date'])
    dfs.append(df_)
    

for chart in tqdm(country_charts3):
    df_ = pd.read_csv(os.path.join('data','chart_data', 'data_202111_202202', chart), 
                      parse_dates=['Date'])
    dfs.append(df_)

ranking_df = pd.concat(dfs,axis=0)    

ranking_df.head()

ranking_df.tail()

ranking_df.info()

ranking_df.shape

ranking_df['URL'].nunique()

### Enrich dataset

country_dict = {'ar':'Argentina', 'at':'Austria', 'au':'Australia', 'be':'Belgium',
       'bo':'Bolivia', 'br':'Brazil', 'ca':'Canada', 'ch':'Switzerland', 
       'cl':'Chile', 'co':'Columbia', 'cr':'CostaRica', 'cz':'CzechRepublic',
       'de':'Germany', 'dk':'Denmark', 'do':'DominicanRepublic', 'ec':'Ecuador', 
       'ee':'Estonia', 'es':'Spain', 'fi':'Finland', 'fr':'France', 
       'gb':'UnitedKingdom', 'global':'World', 'gr':'Greece', 'gt':'Guatemala',
       'hk':'HongKong', 'hn':'Honduras', 'hu':'Hungary', 'id':'Indonesia', 
       'ie':'Ireland', 'is':'Iceland', 'it':'Italy', 'jp':'Japan', 
       'lt':'Lithuania', 'lu':'Luxemborg', 'lv':'Latvia', 'mx':'Mexico', 
       'my':'Malaysia', 'nl':'Netherlands', 'no':'Norway', 'nz':'NewZealand', 
       'pa':'Panama', 'pe':'Peru', 'ph':'Philippines', 'pl':'Poland', 
       'pt':'Portugal', 'py':'Paraguay', 'se':'Sweden', 'sg':'Singapore', 
       'sk':'Slovakia', 'sv':'ElSalvador', 'tr':'Turkey', 'tw':'Taiwan', 'us':'USA', 'uy':'Uruguay',} 

def get_song_id(url):
    
    if type(url)== float:
        return "no id"
    else:
        return url.split('/')[-1]

ranking_df['song_id']= ranking_df['URL'].apply(get_song_id)
ranking_df['Region']= ranking_df['Region'].apply(lambda x: country_dict[x])
ranking_df['release_date']=ranking_df.groupby('song_id')['Date'].transform('min')

ranking_df.head()

ranking_df['Date'].min(), ranking_df['Date'].max()

ranking_df = ranking_df[(ranking_df['song_id']!= 'no id') & (ranking_df['Region']!= 'global') ]

ranking_df.to_csv(os.path.join('data','ranking_all_2017_2022.csv'))

print("That's all folks!!")
