==> 01_get_songs_features_and_genres.ipynb <==
# Extract audio features from each song

from spotipy.oauth2 import SpotifyClientCredentials
import json
import spotipy
import time
import sys

from tqdm.notebook import tqdm
import pandas as pd
import numpy as np
import os
from time import sleep  

SPOTIPY_CLIENT_ID='e1a52feaafdd43ad9eff6dccd21a966e'
SPOTIPY_CLIENT_SECRET='3581d4abb2d9434489ed89cd3160b90e'

client_credentials_manager = SpotifyClientCredentials(client_id=SPOTIPY_CLIENT_ID, client_secret=SPOTIPY_CLIENT_SECRET)
sp = spotipy.Spotify(client_credentials_manager=client_credentials_manager)
sp.trace = False

ranking_features_df= pd.read_csv(os.path.join('raw_data', 'ranking_all_2017_2022.csv'), index_col=0, parse_dates=['Date', 'release_date'])
ranking_features_df.head()

ranking_features_df.info()

## Get songs' audio features

song_features_df= pd.read_csv(os.path.join('raw_data', 'songs_features_2017_2022.csv'), index_col=0)

song_features_df.head()

song_features_df.info()

processed_songs= song_features_df['id'].unique()

target_songs = ranking_features_df['song_id'].unique()
target_songs= set(target_songs)-set(processed_songs)

print(f'Songs to process: {len(target_songs)}')

song_features_lst = []
for song_id in tqdm(target_songs):
    try:
        song_features = sp.audio_features(song_id)[0]
        if song_features:
            song_features_lst.append(song_features)
    except:
        print("Error with song:", song_id)

new_song_features_df = pd.DataFrame(song_features_lst)

new_song_features_df.head()

new_song_features_df.info()

Append new songs' features to existing ones

song_features_df = pd.concat([song_features_df, new_song_features_df], axis=0)

song_features_df.head()

song_features_df.info()

song_features_df.to_csv(os.path.join('raw_data', 'songs_features_2017_2022.csv'))

## Get songs' genres

songs_genres_df= pd.read_csv(os.path.join('raw_data', 'songs_genres_2017_2021.csv'), index_col=0)
songs_genres_df.columns= 'genre_1 genre_2 genre_3 genre_4 genre_5 genre_6 genre_7 genre_8'.split()
songs_genres_df.tail()

processed_songs= songs_genres_df.index

target_songs = ranking_features_df['song_id'].unique()

target_songs= set(target_songs)-set(processed_songs)

all_genres = sp.recommendation_genre_seeds()
all_genres= all_genres['genres']
all_genres += ['rap', 'pop rock', 'trap', 'ranchera', 'corrido', 'cumbia']
all_genres

gender_converter = {'r&b':'r-n-b', 'soundtrack':'soundtracks', 'ost':'soundtracks'}

print("Total songs to be crawled {}".format(len(target_songs)))

song_genres_dict = {}
counter= 0
for song_id in tqdm(target_songs):
    #try:
    track_info= sp.track(song_id)
    track_genres = []
    for artist in track_info['artists']:
        artist_id=artist['id']
        artist_info= sp.artist(artist_id)
        artist_genres = ['hip-hop' if 'hip hop' in g else g for g in artist_info['genres']]
        artist_genres = [gender_converter.get(g,g) for g in artist_genres]
         
        track_genres += artist_genres #artist_info['genres']
        
        
    filtered_genres = [g for g in  track_genres if g in all_genres]
    
    if len(filtered_genres)==0:
        #print(track_genres)
        split_genres = [g.split(' ')[-1] for g in track_genres if len(g.split(' '))>1]
        split_genres += [g.split(' ')[0] for g in track_genres if len(g.split(' '))>1]
        split_genres = [gender_converter.get(g,g) for g in split_genres]
        
        filtered_genres = [g for g in split_genres if g in all_genres]

    if len(filtered_genres)==0:
        print(f'No genres from {track_genres}')
    song_genres_dict[song_id]= set(filtered_genres)
    sleep(0.01)
    
    if counter>= 300:
        print("Update stored songs' genres. ", end="")
        genres_aux = pd.DataFrame.from_dict(song_genres_dict, orient='index')
        rename_cols= {}
        for i in genres_aux.columns:
            rename_cols[i]= 'genre_{}'.format(i+1)
       
        genres_aux= genres_aux.rename(columns=rename_cols)
        for i in range(len(genres_aux.columns),6):
            genres_aux['genre_{}'.format(i+1)]= np.NAN
        
        songs_genres_df = pd.concat([songs_genres_df, genres_aux], axis=0)
        songs_genres_df.to_csv(os.path.join('raw_data', 'songs_genres_2017_2021.csv'))
        song_genres_dict= {}
        counter=0
        print("DONE!")
    counter += 1

genres_aux = pd.DataFrame.from_dict(song_genres_dict, orient='index')
songs_genres_df = pd.concat(songs_genres_df, genres_aux, axis=0)
song_genres_df.to_csv(os.path.join('generated_data', 'songs_genres_2017_2022.csv'))        

print("That's all folks!")
