==> 07b_VAR_prediction_different_clusters.ipynb <==
## Import required libraries

import pandas as pd
import numpy as np
import os

import datetime
from datetime import date, timedelta

#from tqdm.notebook import tqdm
from tqdm import tqdm_notebook

from statsmodels.tsa.vector_ar.vecm import coint_johansen
from statsmodels.tsa.stattools import grangercausalitytests
from statsmodels.tsa.api import VAR
from statsmodels.tsa.stattools import adfuller


from sklearn.preprocessing import MinMaxScaler
from sklearn.metrics import mean_squared_error, mean_absolute_error


import matplotlib.pyplot as plt
%matplotlib inline

import warnings
warnings.filterwarnings("ignore")

#pd.set_option('precision', 3)

## Earth continents

continents = {}
continents['europe']=['Portugal', 'Italy', 'Germany', 'Finland', 'Estonia', 'Hungary', 
                      'Spain', 'Slovakia', 'Ireland', 'Iceland', 'France',  'Norway', 
                      'Lithuania', 'Turkey', 'Switzerland', 'Belgium', 'Denmark',  
                      'Poland', 'Austria', 'Sweden', 'CzechRepublic', 'Netherlands',
                      'UnitedKingdom',  'Latvia', 'Greece', 'Luxemborg']
continents['north_america']= 'USA Canada'.split()
continents['south_america']=['Ecuador', 'Panama', 'Peru',  'Uruguay', 'CostaRica',
                            'Mexico', 'Argentina', 'ElSalvador', 'Chile', 'Brazil', 
                            'Honduras', 'Guatemala','Bolivia', 'Columbia','DominicanRepublic', 'Paraguay']
continents['oceania']='Australia NewZealand'.split()
continents['asia']= ['Indonesia', 'Malaysia', 'Philippines', 'Taiwan', 'Japan', 'Singapore', 'HongKong']

continents['all']=[]
continents['all'].extend(continents['asia'])
continents['all'].extend(continents['oceania'])
continents['all'].extend(continents['south_america'])
continents['all'].extend(continents['north_america'])
continents['all'].extend(continents['europe'])

## Danceability clusters

dance_clust = {}
dance_clust['1']=[ "Portugal", "Italy",    "Finland",  "Estonia",  "Hungary", 
                  "Slovakia", "Ireland",  "Iceland",  "Brazil",   "Canada",  
                  "Belgium",  "Norway",   "Lithuania","New Zealand"    "Turkey",  
                  "USA", "Switzerland"    "Denmark",  "Poland",   "Austria", 
                  "Sweden",   "Czech Republic" "Netherlands",    "UK", 
                  "Australia","Latvia"]

dance_clust['2']= ["Ecuador",     "Panama",      "Germany",     "Spain",      
                   "Peru", "Uruguay",     "Costa Rica",  "Mexico",     
                   "France",      "Argentina",   "El Salvador", "Chile",      
                   "Honduras",    "Guatemala",   "Bolivia",     "Colombia",   
                   "Dominican Republic" "Paraguay",    "Greece",]

dance_clust['3']=[ "Indonesia" ,  "Malaysia" ,   "Philippines", "Taiwan",      "Japan",       "Singapore",   "HongKong"]

## Valence clusters

valence_clust = {}
valence_clust['1']=[ "Portugal",  "Italy","Germany",   "Finland",   "Estonia",  
                    "Hungary",   "Slovakia",  "Ireland",   "Iceland",   "Canada",   
                    "Belgium",   "France",    "Lithuania", "New Zealand"    "Turkey",   
                    "Japan","Switzerland"    "Denmark",   "Poland",    "Austria",  
                    "Sweden",    "Czech Republic", "Netherlands",    "UK",      
                    "Australia", "Latvia",    "Greece"   ]

valence_clust['2']= ["Indonesia","Malaysia", "Norway","Philippines" "Taiwan","USA","Singapore",  "HongKong" ]

valence_clust['3']=["Ecuador", "Panama",  "Spain",   "Peru",   
                    "Brazil",  "Uruguay", "Costa Rica" ,        "Mexico", 
                    "Argentina"  ,        "El Salvador" ,       "Chile",   "Honduras",
                    "Guatemala" ,         "Bolivia", "Colombia","Dominican Republic"
                    "Paraguay"]

## Energy clusters

energy_clust = {}
energy_clust['1']=["Indonesia" ,  "Iceland" ,    "Malaysia",    "Philippines",
                   "Taiwan",      "New Zealand", "USA",        "Singapore","HongKong"   ]

energy_clust['2']= ["Portugal" ,      "Estonia" ,       "Slovakia"  ,     "Ireland" ,       "Canada" ,       
                    "Belgium",        "Mexico" ,        "France"    ,     "Norway" ,        "Lithuania",
                    "Turkey"  ,       "Switzerland"    "Denmark" ,       "Sweden"  ,       "Czech Republic",
                    "Netherlands" ,   "UK"       ,      "World"   ,       "Australia" ,     "Latvia"   ,     
                    "Greece" ]

energy_clust['3']=["Ecuador",    "Panama",     "Italy",      "Germany",                
                   "Finland",    "Hungary",    "Spain",      "Peru",
                   "Brazil",     "Uruguay",    "Costa Rica", "Argentina", 
                   "El Salvador","Chile",      "Japan", "Honduras",  
                   "Guatemala",  "Bolivia",    "Poland",     "Austria",   
                   "Colombia",   "Dominican Republic", "Paraguay"]

resample_freq='W'
train_clust= 'en_3'
eval_clust= 'en_1'

train_countries = energy_clust['3']
eval_countries= energy_clust['2']

ranking_features_df= pd.read_csv(os.path.join('generated_data', 'ranking_features.csv'), index_col=0, parse_dates=['Date', 'release_date'])
ranking_features_df.head()

ranking_features_df.info()

## Read weather data

weather_df= pd.read_csv(os.path.join('generated_data','all_weather.csv'), header=[0,1], index_col=0)
weather_df.index= pd.to_datetime(weather_df.index)
weather_df[('country','World')]= weather_df.mean(axis=1)
weather_df.head()

train_weather= weather_df.iloc[:, weather_df.columns.get_level_values(1).isin(train_countries)].mean(axis=1)
if resample_freq!= 'D':
    train_weather = train_weather.resample(resample_freq).mean()
train_weather.head()

eval_weather= weather_df.iloc[:, weather_df.columns.get_level_values(1).isin(eval_countries)].mean(axis=1)
if resample_freq!= 'D':
    eval_weather = eval_weather.resample(resample_freq).mean()
eval_weather.head()

## Merge all data

def generate_target_multivariate_timeseries(songs_df, country_lst, resample=None):
    
    country_df = ranking_features_df[ranking_features_df['Region'].isin(country_lst)]
    daily_country_df = country_df.drop(columns='Position Streams'.split()).groupby('Date').mean()
 
    if resample != 'D':
        daily_country_df = daily_country_df.resample(resample).mean()
    
    return daily_country_df


train_ts = generate_target_multivariate_timeseries(ranking_features_df, 
                                                    train_countries, 
                                                    resample=resample_freq)

train_ts['weather']= train_weather

train_ts.head()

eval_ts = generate_target_multivariate_timeseries(ranking_features_df, 
                                                    eval_countries, 
                                                    resample=resample_freq)
eval_ts['weather']= eval_weather
eval_ts.head()

# Data visualization

train_ts.head().describe()

### Cointegration test

def cointegration_test(df, alpha=0.05): 
    """Perform Johanson's Cointegration Test and Report Summary"""
    out = coint_johansen(df,-1,7)
    d = {'0.90':0, '0.95':1, '0.99':2}
    traces = out.lr1
    cvts = out.cvt[:, d[str(1-alpha)]]
    
    def adjust(val, length= 6): 
        return str(val).ljust(length)

    # Summary
    print('Name   ::  Test Stat > C(95%)    =>   Signif  \n', '--'*20)
    for col, trace, cvt in zip(df.columns, traces, cvts):
        print(adjust(col), ':: ', adjust(round(trace,2), 9), ">", adjust(cvt, 8), ' =>  ' , trace > cvt)

    print(out.trace_stat)

cointegration_test(train_ts)

### Granger test

maxlag=12
test = 'ssr_chi2test'
def grangers_causation_matrix(data, variables, test='ssr_chi2test', verbose=False):    
    """Check Granger Causality of all possible combinations of the Time series.
    The rows are the response variable, columns are predictors. The values in the table 
    are the P-Values. P-Values lesser than the significance level (0.05), implies 
    the Null Hypothesis that the coefficients of the corresponding past values is 
    zero, that is, the X does not cause Y can be rejected.

    data      : pandas dataframe containing the time series variables
    variables : list containing names of the time series variables.
    """
    df = pd.DataFrame(np.zeros((len(variables), len(variables))), columns=variables, index=variables)
    for c in df.columns:
        for r in df.index:
            test_result = grangercausalitytests(data[[r, c]], maxlag=maxlag, verbose=False)
            p_values = [round(test_result[i+1][0][test][1],4) for i in range(maxlag)]
            if verbose: print(f'Y = {r}, X = {c}, P Values = {p_values}')
            min_p_value = np.min(p_values)
            df.loc[r, c] = min_p_value
    df.columns = [var + '_x' for var in variables]
    df.index = [var + '_y' for var in variables]
    return df

grangers_causation_matrix(train_ts, variables = train_ts.columns)

train_ts

### Augmented Dickey-Fuller Test

def adfuller_test(series, signif=0.05, name='', verbose=False):
    """Perform ADFuller to test for Stationarity of given series and print report"""
    r = adfuller(series, autolag='AIC')
    output = {'test_statistic':round(r[0], 4), 'pvalue':round(r[1], 4), 'n_lags':round(r[2], 4), 'n_obs':r[3]}
    p_value = output['pvalue'] 
    def adjust(val, length= 6): return str(val).ljust(length)

    # Print Summary
    print(f'    Augmented Dickey-Fuller Test on "{name}"', "\n   ", '-'*47)
    print(f' Null Hypothesis: Data has unit root. Non-Stationary.')
    print(f' Significance Level    = {signif}')
    print(f' Test Statistic        = {output["test_statistic"]}')
    print(f' No. Lags Chosen       = {output["n_lags"]}')

    for key,val in r[4].items():
        print(f' Critical value {adjust(key)} = {round(val, 3)}')

    if p_value <= signif:
        print(f" => P-Value = {p_value}. Rejecting Null Hypothesis.")
        print(f" => Series is Stationary.")
    else:
        print(f" => P-Value = {p_value}. Weak evidence to reject the Null Hypothesis.")
        print(f" => Series is Non-Stationary.")

#### Train data

adfuller_test(train_ts['danceability'])

adfuller_test(train_ts['valence'])

adfuller_test(train_ts['energy'])

#### Eval data

adfuller_test(eval_ts['danceability'])

adfuller_test(eval_ts['valence'])

adfuller_test(eval_ts['energy'])

We diferenciate the time series to make it stationary

train_ts_differenced = train_ts.diff().dropna()
diff_=True

adfuller_test(train_ts_differenced['danceability'])

adfuller_test(train_ts_differenced['valence'])

adfuller_test(train_ts_differenced['energy'])

eval_ts_differenced = eval_ts.diff().dropna()
eval_diff_=True

adfuller_test(eval_ts_differenced['danceability'])

adfuller_test(eval_ts_differenced['valence'])

adfuller_test(eval_ts_differenced['energy'])

### Select order for VAR Model

model = VAR(train_ts_differenced)

x = model.select_order(maxlags=24)
x.summary()

n_lags=2

model_fitted = model.fit(n_lags)
model_fitted.summary()

from statsmodels.stats.stattools import durbin_watson
out = durbin_watson(model_fitted.resid)

def adjust(val, length= 6): return str(val).ljust(length)


for col, val in zip(train_ts_differenced.columns, out):
    print(adjust(col), ':', round(val, 2))

### Forecast test

def rolling_window(values_lst, window_size, n_features):
    array= []
    for i in range(values_lst.shape[0]-window_size+1):
        if array is not None:
            array.append(values_lst[i:i+window_size,:].reshape(window_size,n_features))
        else:
            array = values_lst[i:i+window_size,:].reshape(1,window_size,n_features)
    return np.array(array)

def mape_fn(actual, pred): 
    actual, pred = np.array(actual), np.array(pred)
    return np.mean(np.abs((actual - pred) / actual)) * 100


def compute_metrics_as_dataframe_fn(y_valid, y_hat):
    metrics= []
    metrics_global = {'mse':[],'rmse':[],'mae':[],'cvrmse':[],'mape':[],}
    for f in range(y_valid.shape[-1]):
        for t in range(y_valid.shape[1]):
            mae = mean_absolute_error(y_valid[:,t,f], y_hat[:,t,f])
            mse = mean_squared_error(y_valid[:,t,f], y_hat[:,t,f])
            rmse= mean_squared_error(y_valid[:,t,f], y_hat[:,t,f], squared = False)
            cvrmse =  rmse/np.mean(y_valid[:,t,f])*100 # it is a percentage
            mape = mape_fn(y_valid[:,t,f], y_hat[:,t,f])

            metrics.append((f,t, mae, mse, rmse, cvrmse, mape))

    features_map = {0:'danceability', 1:'valence', 2:'energy', 3:'weather'}
    metrics_df = pd.DataFrame.from_records(metrics, columns='feature T MAE MSE RMSE CVRMSE MAPE'.split())
    metrics_df['feature']= metrics_df['feature'].apply(lambda x: features_map[x])
    metrics_df = metrics_df.set_index('T')
    metrics_df.loc['global']= metrics_df.mean(axis=0)
    
    return metrics_df

y_hat = []
n_steps_ahead = 16
nfeatures=4
#n_lags = 8
lr= 0.8 # learning rate

training_size= int(train_ts_differenced.shape[0]* lr)
for i in range(training_size, train_ts_differenced.shape[0]-n_steps_ahead+1):
    X= train_ts_differenced.iloc[:i]
    X_no_diff= train_ts.iloc[1:i+1]
    if not diff_:
        X_no_diff= target_ts.iloc[:i]
    
    Y= eval_ts_differenced.iloc[:i]
    Y_no_diff= eval_ts.iloc[1:i+1]
    if not eval_diff_:
        Y_no_diff= eval_ts.iloc[:i]
    
    model = VAR(X)
    model_fitted = model.fit(n_lags)

    prediction_= model_fitted.forecast(y=Y.values[-n_lags:], steps=n_steps_ahead)
    
    df_forecast = pd.DataFrame(prediction_, 
                               index=eval_ts_differenced.iloc[i:i+n_steps_ahead].index, 
                               columns=eval_ts_differenced.columns + '_1d')

    if diff_:
        columns = eval_ts_differenced.columns
        for col in columns:  
            df_forecast[str(col)] = Y_no_diff[col].iloc[-1] + df_forecast[str(col)+'_1d'].cumsum()
    else:
        for col in columns:  
            df_forecast[str(col)] = df_forecast[str(col)+'_1d']

    #rint(target_ts.iloc[i:i+n_steps_ahead], df_forecast)
    y_hat_lst = df_forecast[columns].values
    y_hat.append(y_hat_lst)


y_true = train_ts.iloc[training_size+1:].values
if not diff_:
    y_true = train_ts.iloc[training_size:].values

y_hat= np.array(y_hat)

var_metrics_df = compute_metrics_as_dataframe_fn(rolling_window(y_true, n_steps_ahead, nfeatures), y_hat)

var_metrics_df

var_metrics_df.T

var_metrics_df[var_metrics_df['feature']=='energy'].T.loc['MAE RMSE CVRMSE'.split()][[0,1,3,7,9,11,15]]

var_metrics_df[var_metrics_df['feature']=='energy'].loc[[0,1,3,7,9,11,15]].mean().round(3)

var_metrics_df.to_csv(os.path.join('prediction_results', 
                                   f'var_metrics_{train_clust}_{eval_clust}_{resample_freq}_diff.csv'))

print("That's all folks!")
